hook up match details gql layer and update UI and nav

+++ core/ui/src/main/kotlin/com/aowen/predcompanion/core/ui/shared/HeroSearchSection.kt
@@ -97,7 +97,7 @@ fun HeroResultCard(
             .fillMaxWidth()
             .clickable {
                 navigateToHeroDetails(
-                    heroDetails.id,
+                    heroDetails.id.toLong(),
                     heroDetails.name
                 )
             }
