+++ feature/heroes/impl/src/main/kotlin/com/aowen/predcompanion/feature/heroes/impl/HeroesScreen.kt
@@ -198,7 +198,7 @@ fun HeroesScreen(
                                 HeroTileCard(
                                     hero = hero,
                                     onClick = {
-                                        navigateToHeroDetails(hero.id, hero.name)
+                                        navigateToHeroDetails(hero.id.toLong(), hero.name)
                                     }
                                 )
                             }
